paid_time_wage needs to be a float

+++ db/migrate/015_add_paid_time_wage_on_timesheets.rb
@@ -1,6 +1,6 @@
 class AddPaidTimeWageOnTimesheets < ActiveRecord::Migration
   def self.up
-    add_column :timesheets, :paid_time_wage, :integer
+    add_column :timesheets, :paid_time_wage, :float
   end
 
   def self.down
